44 physically meaningful configs for control frequency sim frequency discount horizon (#78)
* Calculate discounts from a desired horizon.

Optionally set `discount_horizon` (`GAE_bootstrap_horizon`) instead of `gamma` (`lam`); if it is in the config, then this value will be used in `runner.parse_train_cfg()` to overwrite `gamma` (`lam`) based on the `env.dt`. Unit testing not yet done.

* Calculate discounts from a desired horizon.

Optionally set `discount_horizon` (`GAE_bootstrap_horizon`) instead of `gamma` (`lam`); if it is in the config, then this value will be used in `runner.parse_train_cfg()` to overwrite `gamma` (`lam`) based on the `env.dt`. Unit testing not yet done.

* move discount calculation to task_registry (instead of handling it in the runner)

* fix if/else cases for discount horizons

Co-authored-by: Elijah Stanger-Jones <elijahsj@mit.edu>