Fixed the modification of configs by parameter_dict
Sweeps was not able to modify the settings of train_cfg and env_cfg
before a copy was made respectively in env and the runner, so the
changes were never taking effect. Reworked the call order and some logic
to make it feasible to modify the train_cfg and env_cfg by the wandb
parameters immediately after their creation.